:recycle: #34 refatorado para sincronisar e retornar imoveis
